Getting IPv4 address from domifaddr
Occasionally, the guest IP address is not fetched and Avocado runs fail
during postprocessing with a login timeout. The last failure reports that
no IPv4 DHCP lease was found for the guest MAC address.

This patch obtains the guest IP address using the following command:

virsh domifaddr --full --source arp

If the guest MAC address is found in the command output, its IPv4 address
is obtained and updated in address.cache.

If an IPv4 or IPv6 address is still not found, raise
VMIPAddressMissingError so the existing retry logic attempts to obtain the
address again.

Signed-off-by: Tasmiya Nalatwad <tasmiya@linux.vnet.ibm.com>